+++ app/Campaigns/routes.js
@@ -1,0 +1,11 @@
+const Router = require('express').Router()
+const controller = require('./controller')
+const request = require('./request')
+
+Router.get('/', controller.index)
+Router.post('/', request, controller.store)
+Router.get('/:id', controller.show)
+Router.put('/:id', controller.save)
+Router.delete('/:id', controller.remove)
+
+module.exports = Router